requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pdfplumber==0.11.4
openai==1.58.1
numpy==1.26.4
pinecone-client==5.0.1
pydantic>=2.9,<2.10
python-dotenv==1.0.1
langchain==0.3.13
langchain-community==0.3.11
langchain-google-vertexai==2.0.9
chromadb==0.5.23
python-multipart==0.0.20
langchain-openai==0.2.14
langchain-chroma==0.1.4
pypdf==5.1.0
loguru==0.7.3
azure-identity==1.19.0
google-cloud-secret-manager==2.22.0
PyPDF2==3.0.1

